% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxplot.character}
\alias{boxplot.character}
\alias{boxplot.all}
\alias{boxplot.character.default}
\alias{boxplot.all.default}
\title{Box Plots}
\usage{
boxplot.character(object, ...)

boxplot.all(object, ...)

\method{boxplot.character}{default}(
  object,
  character,
  outliers = TRUE,
  lowerWhisker = 0.05,
  upperWhisker = 0.95,
  col = "white",
  border = "black",
  main = character,
  cex.main = 2,
  xlab = NULL,
  ylab = NULL,
  frame = TRUE,
  pch = 8,
  horizontal = FALSE,
  varwidth = FALSE,
  ...
)

\method{boxplot.all}{default}(
  object,
  folderName = "boxplots",
  outliers = TRUE,
  lowerWhisker = 0.05,
  upperWhisker = 0.95,
  col = "white",
  border = "black",
  main = character,
  cex.main = 2,
  xlab = NULL,
  ylab = NULL,
  frame = TRUE,
  pch = 8,
  horizontal = FALSE,
  varwidth = FALSE,
  ...
)
}
\description{
Box Plots
}
