% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\name{cda.eigenValues}
\alias{cda.eigenValues}
\alias{cda.totalCanonicalStructure}
\title{Export CDA eigenvectors & total canonical structure}
\usage{
cda.eigenValues(object)
cda.totalCanonicalStructure(object)

cda.totalCanonicalStructure(object)
}
\arguments{
\item{object}{an object of class 'cdadata'.}
}
\description{
These functions simplifies exporting eigenvalues and total canonical structure coeficients from 'cdadata' object.
}
\examples{
eVals = cda.eigenValues(myCdaResult)
tcs = cda.totalCanonicalStructure(myCdaResult)

export.res(eVals, file = "./eigenValues.txt")

}
