% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.R
\name{clust}
\alias{clust}
\title{Hierarchical clustering}
\usage{
clust(object, method)

## S3 method for class 'hclust'
plot(x, hang = 0.1, main = "Cluster Dendrogram", ...)
}
\arguments{
\item{object}{an object of class 'morphodata'.}

\item{method}{the agglomeration method to be used: "average" (= UPGMA; default), "complete", "ward.D" (= Ward),
"ward.D2", "single", "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{x}{an object of the type produced by hclust.}

\item{hang}{The fraction of the plot height by which labels should hang below the rest of the plot. A negative value will cause the labels to hang down from 0.}

\item{...}{further arguments to be passed to \code{\link{plot.default}} or other graphical parameters in \code{\link{par}}.}
}
\value{
object of class 'hclust'
}
\description{
Hierarchical cluster analysis of .
}
\details{
This functions performs agglomerative hierarchical clustering based on Euclidean distances.
Typically, populations are used as OTUs. Characters are standardised to a zero mean and a unit standard deviation.

The clustering method should be one of: average (UPGMA; the default), complete or Ward. Note that
the other methods are rarely used in morphometric analyses.

Dendrograms are produced using plot function.
}
\examples{
clustering.UPGMA = clust(data, method = "average")

plot(clustering.UPGMA, cex=0.6, cex.lab=0.8, frame.plot=T, hang=-0.1, main="", xlab="", ylab="distance")

}
