% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{cormat}
\alias{cormat}
\alias{cormat.signifTest}
\title{Correlations of characters}
\usage{
cormat(object, method = "pearson")

cormat.signifTest(object, method = "pearson", alternative = "two.sided")
}
\description{
Correlations of characters
}
