% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.charecter.R
\name{delete.charecter}
\alias{delete.charecter}
\title{Remove morphological characters from an morphodata object}
\usage{
delete.charecter(object, charecter)
}
\arguments{
\item{object}{an object of class 'morphodata'.}

\item{charecter}{vector of characters to be removed.}
}
\value{
object of class 'morphodata'
}
\description{
This function remove morphological characters from all individuals of morphodata object.
}
\examples{
myPartialDataset = delete.charecter(initialdataset, "StemHeight")
myPartialDataset = delete.charecter(initialdataset, c("StemHeight", "LeafNo", "PetalLength"))
}
