% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.population.R
\name{delete.population}
\alias{delete.population}
\title{Remove population from an morphodata object}
\usage{
delete.population(object, populationName)
}
\arguments{
\item{object}{an object of class 'morphodata'.}

\item{populationName}{vector of populations to be removed.}
}
\value{
object of class 'morphodata'
}
\description{
This function remove particular population / populations from morphodata object.
}
\examples{
myPartialDataset = delete.population(initialdataset, "CZLE")
myPartialDataset = delete.population(initialdataset, c("CERV", "DEB", "KOT"))
}
