% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.itemFromMorphodataObject.R
\name{delete.taxon}
\alias{delete.taxon}
\alias{delete.population}
\alias{delete.charecter}
\title{Remove items (Taxa, Populations, morphological characters) from an morphodata object}
\usage{
delete.taxon(object, taxonName)

delete.population(object, populationName)

delete.charecter(object, charecterName)
}
\description{
Remove items (Taxa, Populations, morphological characters) from an morphodata object
}
