% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.points.R
\name{labels.points}
\alias{labels.points}
\title{Add labels to a plot}
\usage{
labels.points(result, axes = c(1,2), ...)

labels.characters(result, axes = c(1,2), ...)
}
\arguments{
\item{...}{further arguments to be passed to \code{\link{text}} or other graphical parameters in \code{\link{par}}.}

\item{result}{an results previously ploted.}

\item{axes}{x, y axes of plot}
}
\description{
This is a generic function for drawing labels to the data points.
}
