% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphodata.R
\name{morphodata}
\alias{morphodata}
\title{Class "morphodata"}
\usage{
## S3 method for class 'morphodata'
summary(x)
}
\arguments{
\item{data}{_____.}

\item{x}{bject of class "morphodata"}
}
\value{
Abject of class "morphodata"
}
\description{
Objects of this class store morphological data.
}
\examples{
read.morphodata("infile.txt")
read.morphodata("clipboard")
}
