\name{nmds.calc}
\alias{nmds.calc}
\title{Non-metric multidimensional scaling (NMDS)}
\usage{
nmds.calc(object, distMethod = "euclidean")
}
\description{
Non-metric multidimensional scaling (NMDS)
}


\details{
Further, multiple runs of the NMDS algorithm are needed to ensure a stable solution has been reached.
# Running the NMDS algorithm multiple times to ensure that the ordination is stable is necessary, as any one run may get "trapped" in local optima which are not representative of true distances. During this process, evaluate the algorithm's results based on how it affects your ecological interpretation of the gradients revealed.


Bray-Curtis coefficient of similarities between avian assemblages and habitat types i

}
