% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.calc.R
\name{pca.calc}
\alias{pca.calc}
\title{Principal component analysis (PCA)}
\usage{
pca.calc(object)
}
\arguments{
\item{object}{an object of class 'morphodata'.}
}
\value{
object of class 'pcadata' containing the following components:

\item{sdev}{the standard deviations of the principal components (i.e., the square roots of the eigenvalues). }
\item{scores}{ordination scores of cases (objects, OTUs), centred and scaled. }
\item{center, scale}{the centering and scaling used, or FALSE.}
\item{eigenVectors}{matrix of eigenvectors (i.e., a matrix of variable loadings).}
\item{eigenValues}{eigenvalues of principal components.}
\item{axesVariance}{eigenvalues as percent. }
\item{cumulativeAxesVariance}{cumulative percentage of eigenvalues. }
}
\description{
This function performs principal component analysis.
}
\details{

}
\examples{
pca.calc(myMorphoData)
}
