% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa.calc.R, R/plot.points.R
\name{pcoa.calc}
\alias{pcoa.calc}
\alias{plotPoints.pcoadata}
\title{Principal coordinates analysis (PCoA)}
\usage{
pcoa.calc(object, distMethod = "euclidean")

\method{plotPoints}{pcoadata}(
  result,
  axes = c(1, 2),
  xlab = NULL,
  ylab = NULL,
  pch = 16,
  col = "black",
  pt.bg = "white",
  breaks = 1,
  ylim = NULL,
  xlim = NULL,
  labels = FALSE,
  legend = FALSE,
  legend.pos = "topright",
  ncol = 1,
  ...
)
}
\description{
Principal coordinates analysis (PCoA)
}
