% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.3Dpoints.R
\name{plot.3Dpoints}
\alias{plot.3Dpoints}
\alias{plot.3Dpoints.pcadata}
\alias{plot.3Dpoints.cdadata}
\title{The default scatterplot 3D function}
\usage{
plot.3Dpoints(result, ...)

\method{plot.3Dpoints}{pcadata}(
  pcaResult,
  axes = c(1, 2, 3),
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  pch = 16,
  col = "black",
  pt.bg = "white",
  phi = 10,
  theta = 2,
  ticktype = "detailed",
  bty = "u",
  type = "n",
  labels = FALSE,
  legend = FALSE,
  legend.pos = "topright",
  ncol = 1,
  ...
)

\method{plot.3Dpoints}{cdadata}(
  cdaResult,
  axes = c(1, 2, 3),
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  pch = 16,
  col = "black",
  pt.bg = "white",
  phi = 10,
  theta = 2,
  ticktype = "detailed",
  bty = "u",
  type = "n",
  labels = FALSE,
  legend = FALSE,
  legend.pos = "topright",
  ncol = 1,
  ...
)
}
\description{
The default scatterplot 3D function
}
