% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.characters.R
\name{plot.characters}
\alias{plot.characters}
\title{Draws eigenvectors as arrows}
\usage{
plot.characters(pcaResult, axes = c(1,2), col = "red", length = 0.1, labels = TRUE, ...)
}
\arguments{
\item{pcaResult}{object of class 'pcadata'.}

\item{...}{further arguments to be passed to \code{\link{arrows}} or other graphical parameters in \code{\link{par}}.}

\item{axes}{x, y axes of plot.}

\item{col}{the color of arro ws.}

\item{length}{length of the edges of the arrow head (in inches).}

\item{labels}{logical, defines if labels are displayed. Only restricted number of parameters are supported. For more precise labels ploting, use \code{\link{labels.characters}} directly.}
}
\description{
The correlation of the characters and ordination axes are visualised as arrows.
}
