% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.legend.R
\name{plot.legend}
\alias{plot.legend}
\alias{plot.legend.pcadata}
\alias{plot.legend.cdadata}
\title{Add legends to plot}
\usage{
plot.legend(result, x = "topright", y = NULL, axes = c(1,2), pch = 16, col = "black", ...)

\method{plot.legend}{pcadata}(
  pcaResult,
  x = "topright",
  y = NULL,
  pch = 16,
  col = "black",
  pt.bg = "white",
  ...
)

\method{plot.legend}{cdadata}(
  cdaResult,
  x = "topright",
  y = NULL,
  pch = 16,
  col = "black",
  pt.bg = "white",
  ...
)
}
\arguments{
\item{result}{an results previously ploted.}

\item{...}{further arguments to be passed to \code{\link{legend}} or other graphical parameters in \code{\link{par}}.}

\item{x, y}{the x and y co-ordinates or a single keyword from the list "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right", and "center", to be used to position the legend.}

\item{pch}{the  plotting symbols of points appearing in the legend.}

\item{col}{the color of points appearing in the legend.}

\item{pt.bg}{the background color for the \code{\link{points}}, corresponding to its argument bg.}
}
\description{
This function can be used to add legends to plots.
}
