% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popul.otu.R
\name{popul.otu}
\alias{popul.otu}
\title{Population means}
\usage{
popul.otu(object)
}
\arguments{
\item{object}{an object of class 'morphodata'.}
}
\value{
object of class 'morphodata'
}
\description{
This function calculates the average value for each character in each population,
with the pairwise deletion of missing data.
}
\details{
This function return morphodata object, where each population is used as the operational
 taxonomic unit, thus is represented by single row with average values for each character.
Note that when using populations as OTUs they are handled with the same weight in all analyses
(disregarding population size, within-population variation, etc.)
}
\examples{
populations = popul.otu(individuals)

}
