\name{qqnormCharacter}
\alias{qqnormCharacter}
\alias{qqnormAll}
\title{Quantile-Quantile Plots}
\usage{
qqnormCharacter(
  object,
  character,
  taxon = NULL,
  qqnorm = TRUE,
  qqline = TRUE,
  ...
)

qqnormAll(
  object,
  folderName = "qqnormPlots",
  taxon = NULL,
  qqnorm = TRUE,
  qqline = TRUE,
  width = 480,
  height = 480,
  units = "px",
  ...
)
}
\description{
Quantile-Quantile Plots
}
