% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.morphodata.R
\name{read.morphodata}
\alias{read.morphodata}
\title{Data Input}
\usage{
read.morphodata(infile, dec=".", ...)
}
\arguments{
\item{infile}{the name of the file which the data are to be read from.}

\item{dec}{the character used in the file for decimal points.}

\item{sep}{the column separator character.}

\item{...}{further arguments to be passed to read.table.}
}
\value{
object of class 'morphodata'
}
\description{
Imports data from tab-delimited text file and produces a 'morphodata' object from it.
}
\details{
The function expect the following data structure:
(1) the first row contains variable names.
(2) the following rows contains individuals, single individual per row;
(3) the first three columns include unique identifiers for individuals, populations and taxa/groups, respectively. Columns have to be named as “ID”, “Population” and “Taxon”;
(4) starting from the fourth column, any number of quantitative or binary morphological characters may be recorded. Any column names can be used (avoiding spaces and special characters);

If there are missing values in the data, they must be represented as empty cells or by the text NA (not quoted), not zero, space or any other character.
}
\examples{
read.morphodata("infile.txt")
read.morphodata("clipboard")
}
