% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cdadata}
\alias{summary.cdadata}
\title{Summarize a structure of a cdadata object.}
\usage{
## S3 method for class 'cdadata'
summary(object)
}
\arguments{
\item{object}{an object of class 'cdadata'.}
}
\description{
This function summarizes the information contained in the a 'cdadata' object.
}
\examples{
summary(object)
}
