% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.morphodata.R
\name{summary.morphodata}
\alias{summary.morphodata}
\title{summary.morphodata}
\usage{
## S3 method for class 'morphodata'
summary(object)
}
\arguments{
\item{object}{an object of class 'morphodata'.}
}
\description{
#' Summarize a structure of morphodata object.
}
\examples{
summary(object)
}
