% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.pcadata}
\alias{summary.pcadata}
\title{Summarize a structure of a pcadata object.}
\usage{
## S3 method for class 'pcadata'
summary(object)
}
\arguments{
\item{object}{an object of class 'pcadata'.}
}
\description{
This function summarizes the information contained in the a pcadata object.
}
\examples{
summary(object)
}
