% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.morphodata}
\alias{summary.pcadata}
\alias{summary.cdadata}
\title{Object Summaries}
\usage{
\method{summary}{morphodata}(object, ...)

\method{summary}{pcadata}(object, ...)

\method{summary}{cdadata}(object, ...)
}
\description{
Object Summaries
}
