% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCharacter.R
\name{transformCharacter}
\alias{transformCharacter}
\title{Transformation of character.}
\usage{
transformCharacter(object, character, FUN, newName = NULL)
}
\description{
Transformation of character.
}
