\name{BoxMTest}
\alias{BoxMTest}
\title{Box's M-test for Homogeneity of Covariance Matrices}
\usage{
BoxMTest(object)
}
\description{
Box's M-test for Homogeneity of Covariance Matrices

The null hypothesis for this test is that the observed covariance matrices for the dependent variables are equal across groups. In other words, a non-significant test result (i.e. one with a large p-value) indicates that the covariance matrices are equal. The generated test statistic is called Box's M statistic.

}
