% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxplot.character}
\alias{boxplot.character}
\alias{boxplot.all}
\title{Box Plots}
\usage{
boxplot.character(object, character, outline = TRUE, lowerWhisker = 0.05, upperWhisker = 0.95, ...)

boxplot.all(object, folderName = "boxplots", ...)
}
\arguments{
\item{object}{an object of class 'morphodata'.}

\item{character}{morphological character used for plotting boxplot}

\item{lowerWhisker}{percentile, to which the lower whisker is extended}

\item{upperWhisker}{percentile, to which the upper whisker is extended}

\item{...}{further arguments to be passed to \code{\link{boxplot}} or \code{\link{bxp}} {graphics}.}

\item{folderName}{the name of folder, where to save produces boxplot}
}
\description{
These functions produce box-and-whisker plot(s) of the given morphological characters.
}
\details{
These functions are wrappers above classicla R boxplot {graphics} function, allowin user to extend whiskers
 to desired percentiles. By default, the  whiskers are extended to the 5\% and 95\% percentiles, because the trimmed range
 (without the most extreme 10% of values) use to be used in taxa descriptions, determination keys, etc. Rectangle define 25th and 75th
 percentiles, bold horizontal line show median.

The \strong{boxplot.all} function produce boxplot for each morphological character and save
 them to folder named in folderName parameter. If it does not exist, new folder is created.
}
\examples{
boxplot.character(myMorphoData, "ST", col = "grey", border = "red")
boxplot.character(myMorphoData, "ST", outline = TRUE, pch = 8, lowerWhisker = 0.05, upperWhisker = 0.95)
boxplot.character(myMorphoData, "ST", outline = FALSE, xlab = "Taxa", ylab = "length", main = "Total stem height (cm)")
boxplot.character(myMorphoData, "ST", varwidth = T, notch = T, boxwex = 0.4, staplewex = 1.3, horizontal = T)

boxplot.all(object, folderName = "boxplots", outline = TRUE, lowerWhisker = 0.05, upperWhisker = 0.95)

}
