% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cda.calc.R
\name{cda.calc}
\alias{cda.calc}
\title{Canonical discriminant analysis (CDA)}
\usage{
cda.calc(object)
}
\arguments{
\item{object}{an object of class 'morphodata'.}
}
\value{
object of class 'cdadata' containing the following components:

\item{totalCanonicalStructure}{totalCanonicalStructuretotalCanonicalStructuretotalCanonicalStructuretotalCanonicalStructure}
}
\description{
This function performs generalized canonical discriminant analysis.
}
\details{

}
\examples{
cdaResult = cda.calc(myMorphoData)

}
