% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cda.calc.R, R/labels.points.R, R/plot.points.R
\name{cda.calc}
\alias{cda.calc}
\alias{labels.points.cdadata}
\alias{plot.points.cdadata}
\title{Canonical discriminant analysis (CDA)}
\usage{
cda.calc(object)

\method{plot.points}{cdadata}(
  cdaResult,
  axes = c(1, 2),
  xlab = NULL,
  ylab = NULL,
  pch = 16,
  col = "black",
  breaks = NULL,
  ylim = NULL,
  bg = "white",
  labels = FALSE,
  legend = FALSE,
  legend.pos = "topright",
  ncol = 1,
  ...
)
}
\arguments{
\item{object}{an object of class 'morphodata'.}
}
\value{
object of class 'cdadata' containing the following components:

\item{totalCanonicalStructure}{totalCanonicalStructuretotalCanonicalStructuretotalCanonicalStructuretotalCanonicalStructure}
}
\description{
This function performs generalized canonical discriminant analysis.
}
\details{

}
\examples{
cdaResult = cda.calc(myMorphoData)

}
