\docType{data}
\name{centaurea}
\alias{centaurea}
\title{25 morphological characters of three species of the \emph{Centaurea phrygia} complex}


\usage{
data(centaurea)
}

\description{
The sample data include portions of data sets from previously published studies by Koutecký (2007) and Koutecký et al. (2012):
25 morphological characters (see the cited studies for details) of the vegetative (stems and leaves) and reproductive structures
(capitula and achenes) of three diploid species of the \emph{Centaurea phrygia} complex: \emph{C. phrygia} L. s.str. (abbreviated "ph"),
\emph{C. pseudophrygia} C.A.Mey. ("ps") and \emph{C. stenolepis} A.Kern. ("st"). Moreover, a fourth group includes the putative hybrid
\emph{C. pseudophrygia} × \emph{C. stenolepis} ("hybr"). The data represent 8, 12, 7 and 6 populations for
each group, respectively, and 20 individuals per population, with one exception in which only 12
individuals were available. All morphological characters are either quantitative (direct measurements,
counts or ratios) or binary (two characters states or presence/absence). 
In four characters of achenes, there are missing data because fruits were not available in all individuals. In two populations of
\emph{C. stenolepis} fruits were completely missing. In total, the data set includes 652 individuals (453
complete) from 33 populations (31 complete).
}

\format{An object of class \code{\link{morphodata}}.
}


\keyword{datasets}

\references{
\strong{Koutecký, P. 2007}. Morphological and ploidy level variation of Centaurea phrygia agg.(Asteraceae) in the Czech Republic, Slovakia and Ukraine. \emph{Folia Geobotanica}, \strong{42}: 77-102.\cr
\strong{Koutecký, P., Štěpánek, J. & Baďurová, T. 2012}. Differentiation between diploid and tetraploid Centaurea phrygia: mating barriers, morphology and geographic distribution. \emph{Preslia}, \strong{84}: 1-32.}
