\name{characters}
\alias{characters}
\title{List Morphological Characters}
\usage{
characters(object)
}

\arguments{

\item{object}{an object of class '\code{\link{morphodata}}', '\code{\link{pcadata}}' or '\code{\link{cdadata}}'.}

}




\description{
Returns list morphological characters of object.
}


\examples{
data(centaurea)

characters(centaurea)

}
