% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif.matrix.R, R/classifda.knn.R,
%   R/classifda.lda.R, R/knn.select.R
\name{classif.matrix}
\alias{classif.matrix}
\alias{classifda.knn}
\alias{classifda.lda}
\alias{knn.select}
\title{Classificatory discriminant analysis}
\usage{
classif.matrix(object, level = "Taxon")

classifda.knn(object, k, crossval = "indiv")

classifda.lda(object, crossval = "indiv")

knn.select(object, crossval = "indiv")
classifda.knn(object, k, crossval = "indiv")

knn.select(object, crossval = "indiv")
}
\arguments{
\item{object}{an object of class 'morphodata'.}

\item{crossval}{crossvalidation mode, sets individual ("indiv"; default) or whole populations ("pop") as leave-out unit.}
}
\description{
These functions computes discriminant function for classifying observations. Linear discriminant function or nonparametric k-nearest-neighbor method can be used.
}
\details{
dsa vrati objekt typu "classifdata
}
\examples{
classifda.lda(morphodata, crossval = "indiv")
classifda.lda(morphodata, crossval = "pop")
classifda.knn(morphodata, k, crossval = "indiv")

}
