\docType{data}
\name{classifdata}
\alias{classifdata}
\title{Class classifdata}

\description{
The classifdata class is designed for storing results of Classificatory discriminant analysis.}

\format{Class \code{classifdata}.
\describe{
	\item{ID}{IDs of each row.}
	\item{Population}{population membership of each row.}
	\item{Taxon}{taxon membership of each row.}

	\item{Classif}{classification from discriminant analysis.}
	
	\item{Classif}{classification from discriminant analysis.}
	
	\item{   }{posterior probabilities of classification into each group (taxon).}
	
	\item{Correct}{logical, correctness of classification.}

	}
}

