% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.taxon.R
\name{delete.taxon}
\alias{delete.taxon}
\title{Remove taxon from an morphodata object}
\usage{
delete.taxon(object, taxonName)
}
\arguments{
\item{object}{an object of class 'morphodata'.}

\item{taxonName}{vector of taxa to be removed.}
}
\value{
object of class 'morphodata'
}
\description{
This function remove particular taxon / taxa from morphodata object.
}
\examples{
myPartialDataset = delete.taxon(initialdataset, "hybr")
myPartialDataset = delete.taxon(initialdataset, c("ph", "st"))
}
