% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.res.R
\name{export.res}
\alias{export.res}
\alias{export.res.morphodata}
\alias{export.res.data.frame}
\alias{export.res.matrix}
\alias{export.res.numeric}
\alias{export.res.classifdata}
\title{Export data}
\usage{
export.res(object, ...)

\method{export.res}{morphodata}(
  object,
  file = "clipboard",
  dec = ".",
  sep = "\\t",
  row.names = F,
  col.names = T
)

\method{export.res}{data.frame}(
  object,
  file = "clipboard",
  dec = ".",
  sep = "\\t",
  row.names = F,
  col.names = T
)

\method{export.res}{matrix}(
  object,
  file = "clipboard",
  dec = ".",
  sep = "\\t",
  row.names = F,
  col.names = T
)

\method{export.res}{numeric}(object, file = "clipboard", dec = ".", sep = "\\t")

\method{export.res}{classifdata}(
  object,
  file = "clipboard",
  dec = ".",
  sep = "\\t",
  row.names = F,
  col.names = T
)
}
\arguments{
\item{object}{an object to be exported.}

\item{...}{further arguments passed to or from other methods.}

\item{file}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}

\item{dec}{the string to use for decimal points in numeric or complex columns: must be a single character.}

\item{sep}{the field separator string. Values within each row of x are separated by this string.}
}
\description{
This is a generic function for exporting results, stored in objects of MorphoTools package.
}
\section{Methods (by class)}{
\itemize{
\item \code{morphodata}: function for exporting 'morphodata' objects

\item \code{data.frame}: function for exporting 'data.frame' objects

\item \code{matrix}: function for exporting 'matrix' objects

\item \code{numeric}: function for exporting 'numeric' objects

\item \code{classifdata}: function for exporting 'classifdata' objects
}}

