% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.calc.R, R/plot.labels.R, R/plot.pcadata.R
\name{pca.calc}
\alias{pca.calc}
\alias{plot.labels.pcadata}
\alias{plot.pcadata}
\title{Principal component analysis (PCA)}
\usage{
pca.calc(object)

## S3 method for class 'pcadata'
plot(pcaResult, axes = c(1,2), pch = 16, col = "black", ...)

## S3 method for class 'pcadata'
plot.labels(pcaResult, axes = c(1,2), ...)
}
\arguments{
\item{object}{an object of class 'morphodata'.}

\item{axes}{x, y axes of plot}

\item{...}{further arguments to be passed to \code{\link{plot.default}}, \code{\link{text}} or other graphical parameters in \code{\link{par}}.}

\item{pch}{a vector of plotting characters or symbols: see \code{\link{points}}.}

\item{col}{the colors for points. Multiple colors can be specified so that each Taxon can be given its own color. If there are fewer colors than points they are recycled in the standard fashion.}

\item{pcaResult}{object of class 'pcadata'}
}
\value{
object of class 'pcadata' containing the following components:

\item{sdev}{the standard deviations of the principal components (i.e., the square roots of the eigenvalues). }
\item{center, scale}{the centering and scaling used, or FALSE.}
\item{objects}{containing ID, Population, Taxon, and ordination scores of cases (objects, OTUs), centred and scaled.}
\item{eigenVectors}{matrix of eigenvectors (i.e., a matrix of variable loadings).}
\item{eigenValues}{eigenvalues of principal components.}
\item{axesVariance}{eigenvalues as percent. }
\item{cumulativeAxesVariance}{cumulative percentage of eigenvalues. }
}
\description{
This function performs principal component analysis.
}
\details{

}
\examples{
pcaResult = pca.calc(myMorphoData)

plot(pcaRes, axes = c(1,2), col = c("red", "green", "blue", "red"), pch = c(20, 17, 8, 21), bg = "orange")
plot(pcaRes, main = "My PCA plot", frame.plot = F, cex = 0.8)

plot.labels(pcaRes, axes = c(1,2), cex = 0.8, col = "red", pos = 4, offset = 0.4)

}
