\name{plot.points}

\alias{plot.points}

\title{The default scatterplot function}

\description{
A generic function for ploting ordination scores stored in '\code{\link{pcadata}}' and '\code{\link{cdadata}}' objects.
}

\usage{
plot.points(object, ...)
}

\arguments{

\item{object}{object of class '\code{\link{pcadata}}' or '\code{\link{cdadata}}'.}

\item{...}{further arguments passed to or from other methods.}
}


