% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.morphodata.R
\name{read.morphodata}
\alias{read.morphodata}
\title{Data Input}
\usage{
read.morphodata(infile, dec=".", ...)
}
\arguments{
\item{infile}{the name of the file which the data are to be read from.}

\item{dec}{the character used in the file for decimal points.}

\item{...}{further arguments to be passed to read.table.}
}
\value{
object of class 'morphodata'
}
\description{
Imports data from tab-delimited text file and produces a 'morphodata' object from it.
}
\examples{
read.morphodata("infile.txt")
read.morphodata("clipboard")
}
