% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.morphodata}
\alias{summary.morphodata}
\title{Summarize a structure of a morphodata object.}
\usage{
\method{summary}{morphodata}(object)
}
\description{
Summarize a structure of a morphodata object.
}
