% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.morphodata}
\alias{summary.morphodata}
\title{Summarize a structure of a morphodata object.}
\usage{
## S3 method for class 'morphodata'
summary(object)
}
\arguments{
\item{object}{an object of class 'morphodata'.}
}
\description{
This function summarizes the information contained in the a 'morphodata' object.
}
\examples{
summary(object)
}
