% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.morphodata.R
\name{summary.morphodata}
\alias{summary.morphodata}
\title{summary.morphodata}
\usage{
summary(x)
}
\arguments{
\item{x}{bject of class "morphodata".}
}
\description{
Summarize morphodata.
}
\examples{
summary(object)
}
