\name{summary}
\alias{summary}
\alias{summary.morphodata}
\alias{summary.pcadata}
\alias{summary.cdadata}

\title{Object Summaries}

\usage{
\method{summary}{morphodata}(object, ...)

\method{summary}{pcadata}(object, ...)

\method{summary}{cdadata}(object, ...)

\method{summary}{classifdata}(object, ...)
}

\arguments{
\item{object}{object of class '\code{\link{pcadata}}', '\code{\link{cdadata}}' or '\code{\link{classifdata}}'.}
\item{...}{additional arguments affecting the summary produced.}

}

\description{
\code{summary} methods for classes "\code{morphodata}", "\code{pcadata}", "\code{cdadata}" and "\code{classifdata}".
}
